webpanel/requirements.txt
Flask
flask-bcrypt
python-dotenv
PyJWT
psycopg2-binary
pyotp
qrcode[pil]
jsonschema
requests
cryptography
gunicorn
psutil
pyquery
APScheduler==3.10.4